requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
boto==2.29.1
dj-database-url==0.3.0
dj-static==0.0.5
django-bootstrap-form==3.1
django-positions==0.5.1
django-relationships==0.3.3
django-storages==1.1.8
django-taggit==0.12
django-toolbelt==0.0.1
django-watson==1.1.3
gunicorn==19.0.0
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
wsgiref==0.1.2
